Factor trades endpoint into `.ib.broker.py`
